api/requirements.txt
fastapi[standard]>=0.114.0,<0.115.0
pydantic>=2.7.0,<3.0.0
uvicorn[standard]==0.30.6
joblib==1.4.2
scikit-learn==1.5.2
transformers==4.44.2
torch==2.4.1
# pytorch==2.3.0
python-arango==8.1.1